libmilter: comparison of integers of different signs
/code/illumos-gate/usr/src/cmd/sendmail/libmilter/comm.c:135:11: error: comparison of integers of different signs: 'mi_int32' (aka 'int') and 'size_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
        if (expl > Maxdatasize)
            ~~~~ ^ ~~~~~~~~~~~
1 error generated.

/code/illumos-gate/usr/src/cmd/sendmail/libmilter/engine.c:1053:33: error: comparison of integers of different signs: 'unsigned long' and 'mi_int32' (aka 'int') [-Werror,-Wsign-compare]
                if ((ctx->ctx_mta_pflags & i) != i)
                     ~~~~~~~~~~~~~~~~~~~~~~~  ^  ~
/code/illumos-gate/usr/src/cmd/sendmail/libmilter/engine.c:1085:32: error: comparison of integers of different signs: 'unsigned long' and 'mi_int32' (aka 'int') [-Werror,-Wsign-compare]
        if ((i & ctx->ctx_mta_aflags) != i)
             ~~~~~~~~~~~~~~~~~~~~~~~  ^  ~
/code/illumos-gate/usr/src/cmd/sendmail/libmilter/engine.c:1095:32: error: comparison of integers of different signs: 'unsigned long' and 'mi_int32' (aka 'int') [-Werror,-Wsign-compare]
        if ((ctx->ctx_mta_pflags & i) != i)
             ~~~~~~~~~~~~~~~~~~~~~~~  ^  ~
/code/illumos-gate/usr/src/cmd/sendmail/libmilter/engine.c:1114:32: error: comparison of integers of different signs: 'unsigned long' and 'mi_int32' (aka 'int') [-Werror,-Wsign-compare]
        if ((ctx->ctx_mta_pflags & i) != i)
             ~~~~~~~~~~~~~~~~~~~~~~~  ^  ~
/code/illumos-gate/usr/src/cmd/sendmail/libmilter/engine.c:1660:8: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
        if (s >= SIZE_NEXT_STATES)
            ~ ^  ~~~~~~~~~~~~~~~~
/code/illumos-gate/usr/src/cmd/sendmail/libmilter/engine.c:1675:9: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
                if (n >= SIZE_NEXT_STATES)
                    ~ ^  ~~~~~~~~~~~~~~~~
/code/illumos-gate/usr/src/cmd/sendmail/libmilter/engine.c:1688:13: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
        } while (s < SIZE_NEXT_STATES);
                 ~ ^ ~~~~~~~~~~~~~~~~
7 errors generated.

+++ usr/src/cmd/sendmail/libmilter/comm.c
@@ -132,7 +132,7 @@ mi_rd_cmd(sd, timeout, cmd, rlen, name)
 	expl = ntohl(expl) - 1;
 	if (expl <= 0)
 		return NULL;
-	if (expl > Maxdatasize)
+	if ((size_t)expl > Maxdatasize)
 	{
 		*cmd = SMFIC_TOOBIG;
 		return NULL;
